Added cheat: Stay in level after star
Added a cheat that makes you stay in a level after collecting a star. You can hold L while collecting a star to temporarily disable this effect. This cheat has no effect on Bowser keys, or stars that already make you stay in the level in the base game.

Also, if you have a better name than "Stay in level after star", be my guest. "Stay in level after getting a star (hold L to cancel)" doesn't fit in the cheat description.